requirements.txt
Django==5.2.7
djangorestframework==3.16.1

channels
channels_redis
daphne

django-wait-for-db==1.0.6
django-cors-headers
djangorestframework-simplejwt

celery>=5.3
django-celery-beat==2.8.0
redis

psycopg[binary]==3.2.4

Pillow

social-auth-app-django
django-allauth>=1.9
dj-rest-auth>=3.1

python-dotenv>=1.0

confluent-kafka>=2.5.0,<3.0

google-auth
google-auth-oauthlib
google-auth-httplib2

django-redis==6.0.0

confluent-kafka
django-redis